+++ src/components/KTable/KTable.vue
@@ -578,7 +578,11 @@ const resizingColumn = ref('')
 const resizerHoveredColumn = ref('')
 // lowest priority - currently hovered resizable column (mouse is somewhere in the <th>)
 const currentHoveredColumn = ref('')
-const hasColumnVisibilityMenu = computed((): boolean => tableHeaders.value.filter((header: TableHeader) => header.hidable).length > 0)
+const hasColumnVisibilityMenu = computed((): boolean => {
+  // has hidable columns, no error/loading/empty state
+  return !!(tableHeaders.value.filter((header: TableHeader) => header.hidable).length > 0 &&
+    !props.hasError && !isTableLoading.value && !props.isLoading && (data.value && data.value.length))
+})
 // columns whose visibility can be toggled
 const visibilityColumns = computed((): TableHeader[] => tableHeaders.value.filter((header: TableHeader) => header.hidable))
 // visibility preferences from the host app (initialized by app)
@@ -707,7 +711,7 @@ const tdlisteners = computed((): any => {
   }
 })
 
-const columnWidths = ref<Record<string, number>>({})
+const columnWidths = ref<Record<string, number>>(props.tablePreferences.columnWidths || {})
 const columnStyles = computed(() => {
   const styles: Record<string, any> = {}
   for (const colKey in columnWidths.value) {