+++ src/components/KToggle/index.ts
@@ -33,7 +33,7 @@ export default defineComponent({
   Example usage:
 
     <KToggle>
-      <button v-slot:default="{isToggled, toggle}" @click="toggle">
+      <button #default="{isToggled, toggle}" @click="toggle">
         {{ isToggled ? 'hello' : 'goodbye' }}
       </button>
       ^^------add slotted content